abstracted projectModal into its own component file in MyProjectsTable
